requirements.txt
faiss_cpu==1.10.0
gradio==5.23.0
groq==0.20.0
numpy==2.2.4
python-dotenv==1.1.0
sentence_transformers==3.4.1
aiofiles==23.2.1
aiohappyeyeballs @ file:///home/conda/feedstock_root/build_artifacts/aiohappyeyeballs_1741775197943/work
aiohttp @ file:///Users/runner/miniforge3/conda-bld/aiohttp_1742268642100/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1734342155601/work
annotated-types==0.7.0
anyio==4.9.0
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1741918516150/work
audioop-lts==0.2.1
Bottleneck @ file:///Users/runner/miniforge3/conda-bld/bottleneck_1729268945324/work
brotlipy @ file:///Users/runner/miniforge3/conda-bld/brotlipy_1725393330668/work
certifi==2025.1.31
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1725560537656/work
charset-normalizer==3.4.1
click==8.1.8
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography-split_1740893626909/work
datasets @ file:///home/conda/feedstock_root/build_artifacts/datasets_1723662581198/work
Deprecated @ file:///home/conda/feedstock_root/build_artifacts/deprecated_1737986966356/work
dill @ file:///home/conda/feedstock_root/build_artifacts/dill_1706434688412/work
distro==1.9.0
docopt==0.6.2
faiss-cpu==1.10.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
frozenlist @ file:///Users/runner/miniforge3/conda-bld/frozenlist_1737645352806/work
fsspec==2025.3.0
gradio==5.23.0
gradio_client==1.8.0
groovy==0.1.2
groq==0.20.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
Jinja2==3.1.6
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1733736026804/work
markdown-it-py==3.0.0
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1733219666454/work
mdurl==0.1.2
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1733302684489/work
multidict @ file:///Users/runner/miniforge3/conda-bld/multidict_1742308168892/work
multiprocess @ file:///Users/runner/miniforge3/conda-bld/multiprocess_1729987579598/work
networkx @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_networkx_1731521053/work
numexpr @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/numexpr_1731706658483/work
numpy==2.2.4
opentelemetry-api @ file:///home/conda/feedstock_root/build_artifacts/opentelemetry-api_1742553828384/work
orjson==3.10.16
packaging==24.2
pandas @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_d6o0iypzho/croot/pandas_1732735102618/work/dist/pandas-2.2.3-cp313-cp313-macosx_10_15_x86_64.whl#sha256=2ef48dcabc48020c306097c5a1b12e25a7ed5e97ad5001b6be3e9b04c16756ef
pillow @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_153gp0xp5x/croot/pillow_1738010255299/work
pipreqs==0.4.13
propcache @ file:///Users/runner/miniforge3/conda-bld/propcache_1737635550915/work
pyarrow @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_38qbz8t7sy/croot/pyarrow_1738239256895/work/python
pyarrow-hotfix @ file:///home/conda/feedstock_root/build_artifacts/pyarrow-hotfix_1734380560621/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1737243356468/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML @ file:///Users/runner/miniforge3/conda-bld/pyyaml_1737454757830/work
regex @ file:///Users/runner/miniforge3/conda-bld/regex_1730952202989/work
requests==2.32.3
rich==13.9.4
ruff==0.11.2
safehttpx==0.1.6
safetensors @ file:///Users/runner/miniforge3/conda-bld/safetensors_1740651472336/work
scikit-learn @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_86qebg7897/croot/scikit-learn_1737988767278/work
scipy @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_06t04e6sdk/croot/scipy_1737122910960/work/dist/scipy-1.15.1-cp313-cp313-macosx_10_15_x86_64.whl#sha256=8d0719e8b2aa562a35f586655b39dec1277a2de1bfc1889140a803574dd1d03a
semantic-version==2.10.0
sentence-transformers @ file:///home/conda/feedstock_root/build_artifacts/sentence-transformers_1738211130290/work
setuptools==70.3.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1736248181614/work
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1741878222898/work
tokenizers @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_72ivn_q5zt/croot/tokenizers_1741371072341/work
tomlkit==0.13.2
torch @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_11jz8t3kkf/croot/libtorch_1738971447389/work
tqdm==4.67.1
transformers @ file:///home/conda/feedstock_root/build_artifacts/transformers_1742577674804/work
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
wheel==0.45.1
wrapt @ file:///Users/runner/miniforge3/conda-bld/wrapt_1736869538973/work
xxhash @ file:///Users/runner/miniforge3/conda-bld/python-xxhash_1740594833887/work
yarg==0.1.10
yarl @ file:///Users/runner/miniforge3/conda-bld/yarl_1737575855520/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
